+++ csg-tutorials/LJ1-LJ2/imc/run.sh
@@ -1,6 +1,0 @@
-#!/bin/sh
-echo 'running csg_inverse --options "settings_pre.xml"'
-csg_inverse --options settings_pre.xml
-rm done
-echo 'running csg_inverse --options "settings.xml"'
-csg_inverse --options settings.xml