+++ csg-tutorials/hexane/atomistic/run_Boltzmann.sh
@@ -1,6 +1,0 @@
-#!/usr/bin/env -S bash -e
-
-#calculate bond and angle distributions with csg_boltzmann
-
-csg_boltzmann --top topol.tpr --trj traj.trr --cg hexane.xml < boltzmann_cmds
-