+++ csg-tutorials/spce/imc/Extract_Energies.sh
@@ -1,1 +1,0 @@
-../../scripts/Extract_Energies.sh