+++ csg-tutorials/urea-water/cibi/pre/run.sh
@@ -1,5 +1,0 @@
-#! /usr/bin/env -S bash -e
-
-echo "Running pre-IBI"
-
-csg_inverse --options settings.xml