+++ csg-tutorials/urea-water/cibi/run.sh
@@ -1,6 +1,0 @@
-#! /usr/bin/env -S bash -e
-
-echo "Running C-IBI"
-
-echo 'running csg_inverse --options "settings.xml"'
-csg_inverse --options settings.xml