+++ csg/src/tools/csg_stat_imc.h
@@ -1,5 +1,5 @@
 /*
- * Copyright 2009-2021 The VOTCA Development Team (http://www.votca.org)
+ * Copyright 2009-2025 The VOTCA Development Team (http://www.votca.org)
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -20,7 +20,7 @@
 
 // VOTCA includes
 #include <votca/tools/average.h>
-#include <votca/tools/histogramnew.h>
+#include <votca/tools/histogram.h>
 #include <votca/tools/property.h>
 
 // Local VOTCA includes
@@ -64,8 +64,8 @@ class Imc {
   struct interaction_t {
     votca::Index index_;
     tools::Property *p_;
-    tools::HistogramNew average_;
-    tools::HistogramNew average_force_;
+    tools::Histogram average_;
+    tools::Histogram average_force_;
     double min_, max_, step_;
     double norm_;
     double cut_;
@@ -137,8 +137,8 @@ class Imc {
 
   class Worker : public CsgApplication::Worker {
    public:
-    std::vector<tools::HistogramNew> current_hists_;
-    std::vector<tools::HistogramNew> current_hists_force_;
+    std::vector<tools::Histogram> current_hists_;
+    std::vector<tools::Histogram> current_hists_force_;
     Imc *imc_;
     double cur_vol_;