+++ xtp/include/votca/xtp/glink.h
@@ -19,6 +19,9 @@
 #ifndef VOTCA_XTP_GLINK_H
 #define VOTCA_XTP_GLINK_H
 
+// Standard includes
+#include <cassert>
+
 // Local VOTCA includes
 #include "eigen.h"