+++ CHANGELOG.rst
@@ -12,6 +12,7 @@ Version 2025.1-dev
 -  Add support for Eigen3 v5.0.0 (#1189)
 -  Remove GMX builds from CI (#1190)
 -  Remove csg_property (#1192)
+-  drop gmx 2019 support from readers (#1193)
 
 Version 2025.1 (released 03.10.25)
 ==================================