+++ CHANGELOG.rst
@@ -14,6 +14,7 @@ Version 2025.1-dev
 -  Remove csg_property (#1192)
 -  drop gmx 2019 support from readers (#1193)
 -  CI: clean up setup action (#1194)
+-  Remove csg_resample (#1195)
 
 Version 2025.1 (released 03.10.25)
 ==================================