+++ CHANGELOG.rst
@@ -15,6 +15,7 @@ Version 2025.1-dev
 -  drop gmx 2019 support from readers (#1193)
 -  CI: clean up setup action (#1194)
 -  Remove csg_resample (#1195)
+-  Move csg_density into csgapps (#1196)
 
 Version 2025.1 (released 03.10.25)
 ==================================